com 6

+++ tasks/seq/vershinina_a_hoare_sort/func_tests/main.cpp
@@ -1,6 +1,7 @@
 #include <gtest/gtest.h>
 
 #include <algorithm>
+#include <cstddef>
 #include <cstdint>
 #include <memory>
 #include <random>
@@ -10,7 +11,7 @@
 #include "seq/vershinina_a_hoare_sort/include/ops_seq.hpp"
 
 namespace {
-std::vector<int> GetRandomVector(size_t len) {
+std::vector<int> GetRandomVector(int len) {
   std::random_device dev;
   std::mt19937 gen(dev());
   std::uniform_int_distribution<> distr(0, 100);
